Revert "History Tab"
This reverts commit 495007f627b2e2df38bee7211d3f1df67c242da3.
